+++ team15duckD2D/player.cpp
@@ -19,8 +19,8 @@ HRESULT player::init()
 	/////////////이건나중에지울거야/////////
 	_tileX = 26;
 	_tileY = 31;
-	_posX = _tileX * 64 + 32;
-	_posY = _tileY * 64 + 32;
+	_posX = (float)_tileX * 64 + 32;
+	_posY = (float)_tileY * 64 + 32;
 	_isMan = true;
 	_state = PS_IDLE_DOWN;
 	_playTime = 5461;
